Returning a Result instead of calling panic!, calling Config::build and handling Errors
We can instead return a Result value that will contain a Config instance
in the successful case and will describe the problem in the error case.

changing the function name from new to build because many programmers
expect new functions to never fail.

unwrap_or_else, which is defined on Result<T, E> by the standard
library.
using unwrap_or_else allows us to define some custom, non-panic! error
handling. If the Result is an Ok value, this method's behavior is
similar to unwrap: it returns the inner value that Ok is wrapping.
However, if the value is an Err value, this method calls the code in the
closure, which is an anonymous function we define and pass as an
argument to unwrap_or_else.

The process::exit function will stop the program immediately and return
the number that was passed as the exit status code.