Working with environment variables
writing a failing test for the case-insensitive search function.
Implementing the search_case_insensitive function.
Adding the ignore_case field that holds a Boolean
Running function to check the ignore_case field's value and use that to
decide whether to call the search function or the
search_case_insensitive function
Checking for the environment variable. The functions for working with
environment variables are in the env module in the standard library.
env::var("some key")
We don't care about the value of the environment variable, just whether
it's set or unset, so we're checking is_ok rather than using unwrap,
expect, or any of the other methods we've seen on Result.